requirements.txt
bzip2==1.0.8
ca-certificates==2024.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
ld_impl_linux-64==2.40
libexpat==2.6.2
libffi==3.4.2
libgcc-ng==13.2.0
libgomp==13.2.0
libnsl==2.0.1
libsqlite==3.45.30
libuuid==2.38.1
libxcrypt==4.4.36
libzlib==1.2.13
ncurses==6.4.20240210
openssl==3.2.1
pip==24.0
pypdf2==3.0.1
python==3.11.9
python-dotenv==1.0.1
readline==8.2
requests==2.31.0
setuptools==69.5.1
tk==8.6.13
tzdata==2024a
urllib3==2.2.1
wheel==0.43.0
xz==5.2.6
